pyproject.toml
[project]
name = "local-nvidia-parkreet-recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "macos-notifications>=0.2.1",
    "nemo-toolkit[asr]>=2.4.0",
    "pynput>=1.8.1",
    "rumps>=0.4.0",
    "sounddevice>=0.5.2",
]
